# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Velozity Project Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
  statusCode: number;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface LoginRequest {
  email: string;
  /** @minLength 6 */
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  project_manager: "project_manager",
  developer: "developer",
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  createdAt: string;
}

export interface LoginResponse {
  accessToken: string;
  user: User;
}

export type CreateUserRequestRole =
  (typeof CreateUserRequestRole)[keyof typeof CreateUserRequestRole];

export const CreateUserRequestRole = {
  admin: "admin",
  project_manager: "project_manager",
  developer: "developer",
} as const;

export interface CreateUserRequest {
  email: string;
  name: string;
  /** @minLength 6 */
  password: string;
  role: CreateUserRequestRole;
}

export type UpdateUserRequestRole =
  (typeof UpdateUserRequestRole)[keyof typeof UpdateUserRequestRole];

export const UpdateUserRequestRole = {
  admin: "admin",
  project_manager: "project_manager",
  developer: "developer",
} as const;

export interface UpdateUserRequest {
  email?: string;
  name?: string;
  role?: UpdateUserRequestRole;
}

export interface Client {
  id: number;
  name: string;
  email: string;
  company: string;
  createdAt: string;
}

export interface CreateClientRequest {
  name: string;
  email: string;
  company: string;
}

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  active: "active",
  completed: "completed",
  on_hold: "on_hold",
  cancelled: "cancelled",
} as const;

export interface Project {
  id: number;
  name: string;
  description?: string;
  status: ProjectStatus;
  clientId: number;
  clientName?: string;
  createdById: number;
  createdByName?: string;
  taskCount: number;
  completedTaskCount: number;
  overdueTaskCount: number;
  createdAt: string;
  updatedAt: string;
}

export type CreateProjectRequestStatus =
  (typeof CreateProjectRequestStatus)[keyof typeof CreateProjectRequestStatus];

export const CreateProjectRequestStatus = {
  active: "active",
  completed: "completed",
  on_hold: "on_hold",
  cancelled: "cancelled",
} as const;

export interface CreateProjectRequest {
  name: string;
  description?: string;
  clientId: number;
  status?: CreateProjectRequestStatus;
}

export type UpdateProjectRequestStatus =
  (typeof UpdateProjectRequestStatus)[keyof typeof UpdateProjectRequestStatus];

export const UpdateProjectRequestStatus = {
  active: "active",
  completed: "completed",
  on_hold: "on_hold",
  cancelled: "cancelled",
} as const;

export interface UpdateProjectRequest {
  name?: string;
  description?: string;
  status?: UpdateProjectRequestStatus;
  clientId?: number;
}

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus];

export const TaskStatus = {
  todo: "todo",
  in_progress: "in_progress",
  in_review: "in_review",
  done: "done",
  overdue: "overdue",
} as const;

export type TaskPriority = (typeof TaskPriority)[keyof typeof TaskPriority];

export const TaskPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface Task {
  id: number;
  title: string;
  description?: string;
  status: TaskStatus;
  priority: TaskPriority;
  dueDate?: string;
  projectId: number;
  projectName?: string;
  assigneeId?: number;
  assigneeName?: string;
  createdById?: number;
  isOverdue: boolean;
  createdAt: string;
  updatedAt: string;
}

export type CreateTaskRequestStatus =
  (typeof CreateTaskRequestStatus)[keyof typeof CreateTaskRequestStatus];

export const CreateTaskRequestStatus = {
  todo: "todo",
  in_progress: "in_progress",
  in_review: "in_review",
  done: "done",
} as const;

export type CreateTaskRequestPriority =
  (typeof CreateTaskRequestPriority)[keyof typeof CreateTaskRequestPriority];

export const CreateTaskRequestPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface CreateTaskRequest {
  title: string;
  description?: string;
  status?: CreateTaskRequestStatus;
  priority: CreateTaskRequestPriority;
  dueDate?: string;
  assigneeId?: number;
}

export type UpdateTaskRequestStatus =
  (typeof UpdateTaskRequestStatus)[keyof typeof UpdateTaskRequestStatus];

export const UpdateTaskRequestStatus = {
  todo: "todo",
  in_progress: "in_progress",
  in_review: "in_review",
  done: "done",
} as const;

export type UpdateTaskRequestPriority =
  (typeof UpdateTaskRequestPriority)[keyof typeof UpdateTaskRequestPriority];

export const UpdateTaskRequestPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface UpdateTaskRequest {
  title?: string;
  description?: string;
  status?: UpdateTaskRequestStatus;
  priority?: UpdateTaskRequestPriority;
  dueDate?: string;
  assigneeId?: number;
}

export interface ActivityEvent {
  id: number;
  taskId: number;
  taskTitle: string;
  projectId: number;
  projectName: string;
  userId: number;
  userName: string;
  action: string;
  fromStatus?: string;
  toStatus?: string;
  description: string;
  createdAt: string;
}

export type NotificationType =
  (typeof NotificationType)[keyof typeof NotificationType];

export const NotificationType = {
  task_assigned: "task_assigned",
  task_in_review: "task_in_review",
  task_status_changed: "task_status_changed",
} as const;

export interface Notification {
  id: number;
  userId: number;
  type: NotificationType;
  title: string;
  message: string;
  taskId?: number;
  projectId?: number;
  isRead: boolean;
  createdAt: string;
}

export interface NotificationsResponse {
  notifications: Notification[];
  unreadCount: number;
}

export type AdminDashboardTasksByStatus = {
  todo: number;
  in_progress: number;
  in_review: number;
  done: number;
  overdue: number;
};

export interface AdminDashboard {
  totalProjects: number;
  totalTasks: number;
  tasksByStatus: AdminDashboardTasksByStatus;
  overdueTaskCount: number;
  activeUsersOnline: number;
}

export type PmDashboardTasksByPriority = {
  low: number;
  medium: number;
  high: number;
  critical: number;
};

export interface PmDashboard {
  projects: Project[];
  tasksByPriority: PmDashboardTasksByPriority;
  upcomingDueDates: Task[];
}

export type DeveloperDashboardTasksByStatus = {
  todo: number;
  in_progress: number;
  in_review: number;
  done: number;
  overdue: number;
};

export interface DeveloperDashboard {
  assignedTasks: Task[];
  tasksByStatus: DeveloperDashboardTasksByStatus;
}

export type ListProjectsParams = {
  status?: string;
  clientId?: number;
};

export type ListProjectTasksParams = {
  status?: string;
  priority?: string;
  dueDateFrom?: string;
  dueDateTo?: string;
  assigneeId?: number;
};

export type ListMyTasksParams = {
  status?: string;
  priority?: string;
  dueDateFrom?: string;
  dueDateTo?: string;
};

export type ListActivityParams = {
  projectId?: number;
  limit?: number;
  before?: string;
};

export type ListNotificationsParams = {
  unreadOnly?: boolean;
};